Fade out after n seconds

+++ BattleArena/include/uniform_buffers.h
@@ -14,7 +14,7 @@ struct UniformBuffer
 	int pointLightCount;
 	int directionalLightCount;
 	int spotlightCount;
-	int pad;
+	float pad;
 };
 
 struct ParticleUniformBuffer
